Changed floats to doubles

+++ src/main/java/machir/emitters/CommonProxy.java
@@ -3,5 +3,5 @@ package machir.emitters;
 import machir.emitters.tileentity.TileEmitter;
 
 public class CommonProxy {
-	public void spawnParticle(TileEmitter tileEmitter, float xOffset, float yOffset, float zOffset, float xVel, float yVel, float zVel) {}
+	public void spawnParticle(TileEmitter tileEmitter, double xOffset, double yOffset, double zOffset, double xVel, double yVel, double zVel) {}
 }
